# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HiFi API",
    "version": "1.0.0",
    "description": "API for products and newsletter subscription"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "tags": [
    { "name": "Products" },
    { "name": "Newsletter" }
  ],
  "paths": {
    "/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get all products",
        "responses": {
          "200": {
            "description": "List of all products",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "name": "Headphones",
                    "price": 199,
                    "image": "http://localhost:4000/images/headphones.jpg"
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/products/{param}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get product by ID or name",
        "parameters": [
          {
            "name": "param",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Product ID or name"
          }
        ],
        "responses": {
          "200": { "description": "Product found" },
          "404": { "description": "Product not found" }
        }
      }
    },
    "/newsletter": {
      "get": {
        "tags": ["Newsletter"],
        "summary": "Get all newsletter emails",
        "responses": {
          "200": { "description": "List of emails" }
        }
      },
      "post": {
        "tags": ["Newsletter"],
        "summary": "Subscribe to newsletter",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "email": { "type": "string" } }
              },
              "example": { "email": "jens@jensen.dk" }
            }
          }
        },
        "responses": {
          "201": { "description": "Subscribed successfully" },
          "400": { "description": "Invalid email or already subscribed" }
        }
      }
    }
  }
}